Tried to resolve some warnings

+++ src/TableOfContents.h
@@ -120,8 +120,8 @@ private:
   unsigned int m_numberOfCaptionsDragged;
   GroupCell *m_cellRightClickedOn;
   //! The item that was dragged away at the start of the current drag-and-drop
-  size_t m_dragStart = -1;
-  size_t m_dragStop = -1;
+  long m_dragStart = -1;
+  long m_dragStop = -1;
   long m_dragCurrentPos = -1;
   //! The position the dragged item was when we last displayed the reordered toc
   int m_dragFeedback_Last = -1;
